refactor(rename to userData): rename BerH Anwaltliche Vertretung
