+++ app/domains/beratungshilfe/formular/context.ts
@@ -1,6 +1,6 @@
 import { dokumenteContext } from "~/domains/beratungshilfe/formular/abgabe/dokumente/context";
 import { beratungshilfePersoenlicheDaten } from "~/domains/beratungshilfe/formular/persoenlicheDaten/context";
-import { beratungshilfeAnwaltlicheVertretung } from "./anwaltlicheVertretung/context";
+import { beratungshilfeAnwaltlicheVertretung } from "./anwaltlicheVertretung/userData";
 import { beratungshilfeFinanzielleAngaben } from "./finanzielleAngaben/context";
 import { beratungshilfeGrundvoraussetzungen } from "./grundvoraussetzung/context";
 import { beratungshilfeRechtsproblem } from "./rechtsproblem/context";
